Don't trim redirect_path if path is just /
The redirect_path had a trailing / trimmed, even when the path was
just /. This broke code_key detection on line 133 because the
concatenated path didn't match the parsed_auth_response key.

Some organisations prevent a user from registering their own app, and
as such specifying their own redirect URI. For organisations that have
configure Thunderbird the redirect URI is localhost on any port with a
path of /. It's possible users in this situation may want to use a
path of just /, hence this fix.

Signed-off-by: Laurence Rochfort <laurence.rochfort@gmail.com>